Extract regex expansion into utils package
